Fix content build exception: System.Reflection.ReflectionTypeLoadException: Unable to load one or more of the requested types.
This seems to be because some monogame dependencies are not present. Although not used by monogame the are looked for during the reflective type discovery of content writers.
The assembly load order matters. If these are hit before your assembly that has your Asset serialization type then you'll get the exception above.